deprecate old main widget in favor of new
